+++ android-interop-testing/build.gradle
@@ -73,7 +73,6 @@ dependencies {
             project(':grpc-protobuf-lite'),
             project(':grpc-stub'),
             project(':grpc-testing'),
-            libraries.hdrhistogram,
             libraries.junit,
             libraries.truth,
             libraries.androidx.test.rules,
